+++ src/ssh/HISTORY.md
@@ -1,6 +1,14 @@
 Release History
 ===============
 
+0.1.2
+-----
+* Add support for hardware tokens (don't require the private key be passed in)
+* Add support for cert signing only
+* Add numerous short parameters
+* Add support for and switch default behavior to ephemeral keypair generation when no public key is passed in
+* Add support for Host * by not writing Hostname into ssh_config files
+
 0.1.1
 -----
 * Fix bash not work problem.